+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   *  `Get Global Emotes`
   *  `Get Emote Sets`
   *  `Get Channel Stream Schedule`
+  *  `Update Channel Stream Schedule`
 * Added Channel Terms to pubsub `chat_moderator_actions`
 * Added `user-moderation-notifications` topic to pubsub
 * Added `extendsub` to pubsub `channel-subscribe-events-v1`